+++ Sources/EudiWalletKit/Services/StorageManager.swift
@@ -24,25 +24,19 @@ import CryptoKit
 /// Storage manager. Provides services and view models
 public class StorageManager: ObservableObject {
 	public static let knownDocTypes = [EuPidModel.euPidDocType, IsoMdlModel.isoDocType]
-	/// Array of doc.types of documents loaded in the wallet
-	public var docTypes: [String] { mdocModels.map(\.docType) }
 	/// Array of document models loaded in the wallet
-	@Published public var mdocModels: [any MdocDecodable] = []
-	/// Array of document identifiers loaded in the wallet
-	public var documentIds: [String] { mdocModels.map(\.id) }
+	@Published public private(set) var mdocModels: [any MdocDecodable] = []
 	var storageService: any DataStorageService
 	/// Whether wallet currently has loaded data
-	@Published public var hasData: Bool = false
+	@Published public private(set) var hasData: Bool = false
 	/// Whether wallet currently has loaded a document with doc.type included in the ``knownDocTypes`` array
-	@Published public var hasWellKnownData: Bool = false
+	@Published public private(set) var hasWellKnownData: Bool = false
 	/// Count of documents loaded in the wallet
-	@Published public var docCount: Int = 0
+	@Published public private(set) var docCount: Int = 0
 	/// The first driver license model loaded in the wallet (deprecated)
-	@Published public var mdlModel: IsoMdlModel?
+	@Published public private(set) var mdlModel: IsoMdlModel?
 	/// The first PID model loaded in the wallet (deprecated)
-	@Published public var pidModel: EuPidModel?
-	/// Other document models loaded in the wallet
-	@Published public var otherModels: [GenericMdocModel] = []
+	@Published public private(set) var pidModel: EuPidModel?
 	/// Error object with localized message
 	@Published public var uiError: WalletError?
 	let logger: Logger
@@ -55,11 +49,10 @@ public class StorageManager: ObservableObject {
 	@MainActor
 	func refreshPublishedVars() {
 		hasData = mdocModels.count > 0
-		hasWellKnownData = hasData && !Set(docTypes).isDisjoint(with: Self.knownDocTypes)
+		hasWellKnownData = hasData && !Set(mdocModels.map(\.docType)).isDisjoint(with: Self.knownDocTypes)
 		docCount = mdocModels.count
 		mdlModel = getTypedDoc()
 		pidModel = getTypedDoc()
-		otherModels = getTypedDocs()
 	}
 	
 	@MainActor
@@ -116,7 +109,7 @@ public class StorageManager: ObservableObject {
 	/// Get document model by index
 	/// - Parameter index: Index in array of loaded models
 	/// - Returns: The ``MdocDecodable`` model
-	public func getDocumentModel(index: Int) -> (any MdocDecodable)? {
+	func getDocumentModel(index: Int) -> (any MdocDecodable)? {
 		guard index < mdocModels.count else { return nil }
 		return mdocModels[index]
 	}
@@ -125,7 +118,7 @@ public class StorageManager: ObservableObject {
 	/// - Parameter id: The id of the document model to return
 	/// - Returns: The ``MdocDecodable`` model
 	public func getDocumentModel(id: String) ->  (any MdocDecodable)? {
-		guard let i = documentIds.firstIndex(of: id)  else { return nil }
+		guard let i = mdocModels.map(\.id).firstIndex(of: id)  else { return nil }
 		return getDocumentModel(index: i)
 	}
 	
@@ -133,8 +126,8 @@ public class StorageManager: ObservableObject {
 	/// - Parameter docType: The docType of the document model to return
 	/// - Returns: The ``MdocDecodable`` model
 	public func getDocumentModels(docType: String) -> [any MdocDecodable] {
-		return (0..<docTypes.count).compactMap { i in
-			guard docTypes[i] == docType else { return nil }
+		return (0..<mdocModels.count).compactMap { i in
+			guard mdocModels[i].docType == docType else { return nil }
 			return getDocumentModel(index: i)
 		}
 	}
@@ -142,7 +135,7 @@ public class StorageManager: ObservableObject {
 	/// Delete document by id
 	/// - Parameter id: Document id
 	public func deleteDocument(id: String) async throws {
-		guard let i: Array<String?>.Index = documentIds.firstIndex(of: id)  else { return }
+		guard let i: Array<String?>.Index = mdocModels.map(\.id).firstIndex(of: id)  else { return }
 		do {
 			try await deleteDocument(index: i)
 		} catch {
@@ -152,14 +145,14 @@ public class StorageManager: ObservableObject {
 	}
 	/// Delete document by Index
 	/// - Parameter index: Index in array of loaded models
-	public func deleteDocument(index: Int) async throws {
-		guard index < documentIds.count else { return }
+	func deleteDocument(index: Int) async throws {
+		guard index < mdocModels.count else { return }
 		let id = mdocModels[index].id
 		do {
 			try storageService.deleteDocument(id: id, status: .issued)
 			await MainActor.run {
-				if docTypes[index] == IsoMdlModel.isoDocType { mdlModel = nil }
-				if docTypes[index] == EuPidModel.euPidDocType { pidModel = nil }
+				if mdocModels[index].docType == IsoMdlModel.isoDocType { mdlModel = nil }
+				if mdocModels[index].docType == EuPidModel.euPidDocType { pidModel = nil }
 				mdocModels.remove(at: index)
 			}
 			await refreshPublishedVars()